api/requirements.txt
flask
psycopg2
SQLAlchemy
python-dotenv